requirements.txt
textual>=0.50.0
websockets>=12.0
cryptography>=41.0.7
argon2-cffi>=23.1.0
rich>=13.7.1
